Sorting task 0 error

+++ 0x1E-search_algorithms/0-linear.c
@@ -11,14 +11,14 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-        if (array == NULL)
-                return (-1);
+	if (array == NULL)
+		return (-1);
 
-        for (size_t i = 0; i < size; i++)
-        {
-                printf("Value checked array[%lu] = [%d]\n", i, array[i]);
-                if (value == array[i])
-                        return ((int)i);
-        }
-        return (-1);
+	for (size_t i = 0; i < size; i++)
+	{
+		printf("Value checked array[%lu] = [%d]\n", i, array[i]);
+		if (value == array[i])
+			return (i);
+	}
+	return (-1);
 }
